% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiomassByLandType.R
\name{calcBiomassByLandType}
\alias{calcBiomassByLandType}
\title{Calculate Biomass by land type}
\usage{
calcBiomassByLandType(subtype)
}
\arguments{
\item{subtype}{"aboveground" or "belowground"}
}
\value{
data
}
\description{
Estimate soil organic carbon content split by land type.
}
\examples{
\dontrun{
a <- calcOutput("BiomassbyLandType", subtype = "aboveground")
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
