% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEcoregions2017Raster.R
\name{calcEcoregions2017Raster}
\alias{calcEcoregions2017Raster}
\title{calcEcoregions2017Raster}
\usage{
calcEcoregions2017Raster(
  ecoregionsColumn = "ECO_BIOME_",
  nrows = 720,
  ncols = 1440
)
}
\arguments{
\item{ecoregionsColumn}{Which column of the ecoregions dataset should be in the raster.}

\item{nrows}{Number of rows in the raster produced.}

\item{ncols}{Number of columns in the raster produced.}
}
\value{
A raster with ecoregion information (ecoregionsColumn) as a magclass object
}
\description{
Produces a raster with ecoregion information as a magclass object.
}
\author{
Pascal Führlich
}
