% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTypeAreasAggregated.R
\name{calcLandTypeAreasAggregated}
\alias{calcLandTypeAreasAggregated}
\title{Aggregate land type areas}
\usage{
calcLandTypeAreasAggregated(res, zeroValue = 0)
}
\arguments{
\item{res}{target resolution}

\item{zeroValue}{value for cells with no area. Useful to change to a very
small number if the area information should be used as weight.}
}
\value{
data
}
\description{
Aggregate land type area information to a desired resolution
}
\examples{
\dontrun{
a <- calcOutput("LandTypeAreas")
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
