% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWorldCountries.R
\name{calcWorldCountries}
\alias{calcWorldCountries}
\title{Calculate world countries}
\usage{
calcWorldCountries()
}
\value{
data
}
\description{
Partition the world into countries with their respective ISO codes
}
\examples{
\dontrun{
a <- calcOutput("WorldCountries", aggregate = FALSE)
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Patrick v. Jeetze, Jan Philipp Dietrich
}
