% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGSOCseq.R
\name{convertGSOCseq}
\alias{convertGSOCseq}
\title{Convert GSOCseq}
\usage{
convertGSOCseq(x)
}
\arguments{
\item{x}{data}
}
\value{
data
}
\description{
Convert GSOCseq data
}
\author{
Jan Philipp Dietrich
}
