% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullORGANIC.R
\name{fullORGANIC}
\alias{fullORGANIC}
\title{fullORGANIC}
\usage{
fullORGANIC(rev)
}
\arguments{
\item{rev}{data revision}
}
\description{
Bundle function for mrorganic which bundles all relevant outputs and write
them to a tgz file.
}
\details{
Use \code{regionmapping = "regionmappingGTAP11.csv")} to produce outputs
for GTAP11 world regions.
}
\examples{
\dontrun{
retrieveData("Organic", rev = 1, regionmapping = "regionmappingGTAP11.csv")
}

}
\seealso{
\code{\link{retrieveData}}, \code{\link{calcOutput}}, \code{\link{readSource}},
\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
