% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrorganic-package.R
\docType{package}
\name{mrorganic-package}
\alias{mrorganic}
\alias{mrorganic-package}
\title{mrorganic: Provide SOC and biomass data}
\description{
mrorganic greatly simplifies downloading and preparing SOC and biomass data for further scientific work.
}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}
}

}
