% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readESACCI.R
\name{readESACCI}
\alias{readESACCI}
\title{Read ESACCI}
\usage{
readESACCI(subtype = "landcover2010")
}
\arguments{
\item{subtype}{Subtype to be read in}
}
\value{
data
}
\description{
Read ESACCI data
}
\examples{
\dontrun{
a <- readSource("ESACCI")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
