% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddEcoregions.R
\name{toolAddEcoregions}
\alias{toolAddEcoregions}
\title{toolAddEcoregions}
\usage{
toolAddEcoregions(
  x,
  ecoregionsColumns = c("ECO_BIOME_", "BIOME_NAME", "REALM")
)
}
\arguments{
\item{x}{A SpatRaster}

\item{ecoregionsColumns}{Which columns of the ecoregions dataset should be added.}
}
\value{
x with additional ecoregion information
}
\description{
Enriches rasterdata with ecoregion information.
}
\author{
Pascal Führlich
}
