% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregatebyLandType.R
\name{toolAggregateByLandType}
\alias{toolAggregateByLandType}
\title{toolAggregateByLandType}
\usage{
toolAggregateByLandType(x, weight)
}
\arguments{
\item{x}{data to be aggregated}

\item{weight}{weight to be used for aggregation (land type)}
}
\value{
data
}
\description{
Aggregate raster object by land type and return as 0.25degx0.25deg magclass
object
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
