% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTypeAreas.R
\name{calcLandTypeAreas}
\alias{calcLandTypeAreas}
\title{Calculate land type areas}
\usage{
calcLandTypeAreas(categories = "CropGrassOther")
}
\arguments{
\item{categories}{Categories land should be segregated into. Only available
options are currently "CropGrassOther" and "CropGrassForestOthervegResidual".}
}
\value{
data
}
\description{
Convert ESACCI landcover information into area information for selected
land types cropland, grassland and other.
}
\examples{
\dontrun{
a <- calcOutput("LandTypeAreas")
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
