% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOCbyLandType.R
\name{calcSOCbyLandType}
\alias{calcSOCbyLandType}
\title{Calculate SOC by land type}
\usage{
calcSOCbyLandType()
}
\value{
data
}
\description{
Estimate soil organic carbon content split by land type
}
\examples{
\dontrun{
a <- calcOutput("SOCbyLandType")
}
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
