% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullORGANIC.R
\name{fullORGANIC}
\alias{fullORGANIC}
\title{fullORGANIC}
\usage{
fullORGANIC()
}
\description{
Bundle function for mrorganic which bundles all relevant outputs and write
them to a tgz file.
}
\details{
Use \code{regionmapping = "regionmappingGTAP11.csv")} to produce outputs
for GTAP11 world regions.
}
\examples{
\dontrun{
retrieveData("Organic", regionmapping = "regionmappingGTAP11.csv")
}

}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
