% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrorganic-package.R
\docType{package}
\name{mrorganic-package}
\alias{mrorganic}
\alias{mrorganic-package}
\title{mrorganic: madrat based MAgPIE Input Data Library}
\description{
Provides functions for MAgPIE country and cellular input data generation.
}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}
}

}
