% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrorganic-package.R
\docType{package}
\name{mrorganic-package}
\alias{mrorganic}
\alias{mrorganic-package}
\title{mrorganic: Provide soil organic carbon and biomass data}
\description{
mrorganic supports downloading and preparing soil organic carbon and biomass data for further scientific work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tscheypidi/mrorganic}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}
}

Other contributors:
\itemize{
  \item Hugo Valin [contributor]
}

}
