% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(x, name, createPng = FALSE, ...)
}
\arguments{
\item{x}{The data to plot as magclass object}

\item{name}{The name/title of the plot/png}

\item{createPng}{If TRUE save plot as png}

\item{...}{Further arguments passed to terra::plot}
}
\description{
Create a plot showing a map.
}
