% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGSOCseq.R
\name{readGSOCseq}
\alias{readGSOCseq}
\title{Read GSOCseq}
\usage{
readGSOCseq(subtype = "ini")
}
\arguments{
\item{subtype}{data set to use (currently available "ini" and "finalSSM1")}
}
\value{
data
}
\description{
Read GSOCseq data
}
\examples{
\dontrun{
a <- readSource("GSOCseq")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
