% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldBankMaps.R
\name{readWorldBankMaps}
\alias{readWorldBankMaps}
\title{readWorldBankMaps}
\usage{
readWorldBankMaps(subtype = "CountryBoundaries")
}
\arguments{
\item{subtype}{data set to use (currently available "CountryBoundaries" and "DisputedAreasBoundaries")}
}
\value{
data
}
\description{
Read World Bank Maps
}
\examples{
\dontrun{
a <- readSource("WorldBankMaps")
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
