% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddEcoregions.R
\name{toolAddEcoregions}
\alias{toolAddEcoregions}
\title{toolAddEcoregions}
\usage{
toolAddEcoregions(x, ecoregions, regionmapping)
}
\arguments{
\item{x}{A magclass object}

\item{ecoregions}{Ecoregions dataframe, as returned by calcEcoregions2017Raster}

\item{regionmapping}{A regionmapping}
}
\value{
x with additional ecoregion and region information. Spatial subdimensions:
        c("x", "y", "country", "region", "ECO_BIOME_", "country_ECO_BIOME_", "region_ECO_BIOME_")
}
\description{
Enriches a magclass object with ecoregion and region information.
}
\author{
Pascal Führlich
}
